refactor: moved navbar component into layout directory
